Fix short-circuiting logic for validation
When a validation is being done with a context, it makes sense to
validate all items, so that the user of the library can see all the
items that failed the constraints, so the logic should not
short-circuit.

When the validation is being done without a context on the other hand,
it can (and should) be short-circuited to improve performance.