Fix default value behavior
The AVRO specification defines that the default value should be used by
the reader side only, to resolve a value for a field that is not present
in the writer schema. The writer should not use the default value.

There was also an issue that would cause the default to be used by the
reader, in case the serialized value is a null, and the default value is
not null.